Add files via upload
1. Introduction
This project designs and implements a simplified storage management system. The system consists a number of agents and a manager. An agent is a piece of software that runs on a computer and manages the storage device such as redundant array of independent disk (RAID), which is connected to the computer. The manager is another piece of software that runs by a system administrator who manages the storage devices across the enterprise. Each agent periodically reports the system heath information to the manager and executes the commands from the manager. 
2. System design
2.1. Agent
The agent starts by launching two threads, BeaconSender and CmdAgent. 
BeaconSender sends a UDP packet, referred to as a beacon, periodically (e.g., every 1 minute) to the manager. 
CmdAgent listens to a port that the manager sends commands through TCP.  When receiving a command, CmdAgent executes the command and returns the result to the manager. 
2.2. Manager
The manager starts by launch